+++ src/headless/shared/api/presence.js
@@ -10,8 +10,8 @@ export default {
         /**
          * Send out a presence stanza
          * @method _converse.api.user.presence.send
-         * @param { String } type
-         * @param { String } to
+         * @param { String } [type]
+         * @param { String } [to]
          * @param { String } [status] - An optional status message
          * @param { Array<Element>|Array<Strophe.Builder>|Element|Strophe.Builder } [child_nodes]
          *  Nodes(s) to be added as child nodes of the `presence` XML element.